Moved conversion of vitess bindings into engine, from handler. New test methods to deal with same
