+++ engine.go
@@ -17,8 +17,12 @@ package sqle
 import (
 	"fmt"
 	"os"
+	"strconv"
 	"sync"
 
+	"github.com/dolthub/go-mysql-server/sql/types"
+	"github.com/dolthub/vitess/go/sqltypes"
+	"github.com/dolthub/vitess/go/vt/proto/query"
 	"github.com/pkg/errors"
 
 	"github.com/dolthub/go-mysql-server/memory"
@@ -247,6 +251,127 @@ func (e *Engine) Query(ctx *sql.Context, query string) (sql.Schema, sql.RowIter,
 // QueryWithBindings executes the query given with the bindings provided
 func (e *Engine) QueryWithBindings(ctx *sql.Context, query string, bindings map[string]sql.Expression) (sql.Schema, sql.RowIter, error) {
 	return e.QueryNodeWithBindings(ctx, query, nil, bindings)
+}
+
+// QueryNodeWithVitessBindings executes the query given with the bindings provided
+func (e *Engine) QueryNodeWithVitessBindings(ctx *sql.Context, query string, parsed sql.Node, bindings map[string]*query.BindVariable) (sql.Schema, sql.RowIter, error) {
+	var sqlBindings map[string]sql.Expression
+	if len(bindings) > 0 {
+		var err error
+		sqlBindings, err = bindingsToExprs(bindings)
+		if err != nil {
+			ctx.GetLogger().WithError(err).Errorf("Error processing bindings")
+			return nil, nil, err
+		}
+	}
+	
+	return e.QueryNodeWithBindings(ctx, query, parsed, sqlBindings)
+}
+
+func bindingsToExprs(bindings map[string]*query.BindVariable) (map[string]sql.Expression, error) {
+	res := make(map[string]sql.Expression, len(bindings))
+	for k, v := range bindings {
+		v, err := sqltypes.NewValue(v.Type, v.Value)
+		if err != nil {
+			return nil, err
+		}
+		switch {
+		case v.Type() == sqltypes.Year:
+			v, _, err := types.Year.Convert(string(v.ToBytes()))
+			if err != nil {
+				return nil, err
+			}
+			res[k] = expression.NewLiteral(v, types.Year)
+		case sqltypes.IsSigned(v.Type()):
+			v, err := strconv.ParseInt(string(v.ToBytes()), 0, 64)
+			if err != nil {
+				return nil, err
+			}
+			t := types.Int64
+			c, _, err := t.Convert(v)
+			if err != nil {
+				return nil, err
+			}
+			res[k] = expression.NewLiteral(c, t)
+		case sqltypes.IsUnsigned(v.Type()):
+			v, err := strconv.ParseUint(string(v.ToBytes()), 0, 64)
+			if err != nil {
+				return nil, err
+			}
+			t := types.Uint64
+			c, _, err := t.Convert(v)
+			if err != nil {
+				return nil, err
+			}
+			res[k] = expression.NewLiteral(c, t)
+		case sqltypes.IsFloat(v.Type()):
+			v, err := strconv.ParseFloat(string(v.ToBytes()), 64)
+			if err != nil {
+				return nil, err
+			}
+			t := types.Float64
+			c, _, err := t.Convert(v)
+			if err != nil {
+				return nil, err
+			}
+			res[k] = expression.NewLiteral(c, t)
+		case v.Type() == sqltypes.Decimal:
+			v, _, err := types.InternalDecimalType.Convert(string(v.ToBytes()))
+			if err != nil {
+				return nil, err
+			}
+			res[k] = expression.NewLiteral(v, types.InternalDecimalType)
+		case v.Type() == sqltypes.Bit:
+			t := types.MustCreateBitType(types.BitTypeMaxBits)
+			v, _, err := t.Convert(v.ToBytes())
+			if err != nil {
+				return nil, err
+			}
+			res[k] = expression.NewLiteral(v, t)
+		case v.Type() == sqltypes.Null:
+			res[k] = expression.NewLiteral(nil, types.Null)
+		case v.Type() == sqltypes.Blob || v.Type() == sqltypes.VarBinary || v.Type() == sqltypes.Binary:
+			t, err := types.CreateBinary(v.Type(), int64(len(v.ToBytes())))
+			if err != nil {
+				return nil, err
+			}
+			v, _, err := t.Convert(v.ToBytes())
+			if err != nil {
+				return nil, err
+			}
+			res[k] = expression.NewLiteral(v, t)
+		case v.Type() == sqltypes.Text || v.Type() == sqltypes.VarChar || v.Type() == sqltypes.Char:
+			t, err := types.CreateStringWithDefaults(v.Type(), int64(len(v.ToBytes())))
+			if err != nil {
+				return nil, err
+			}
+			v, _, err := t.Convert(v.ToBytes())
+			if err != nil {
+				return nil, err
+			}
+			res[k] = expression.NewLiteral(v, t)
+		case v.Type() == sqltypes.Date || v.Type() == sqltypes.Datetime || v.Type() == sqltypes.Timestamp:
+			t, err := types.CreateDatetimeType(v.Type())
+			if err != nil {
+				return nil, err
+			}
+			v, _, err := t.Convert(string(v.ToBytes()))
+			if err != nil {
+				return nil, err
+			}
+			res[k] = expression.NewLiteral(v, t)
+		case v.Type() == sqltypes.Time:
+			t := types.Time
+			v, _, err := t.Convert(string(v.ToBytes()))
+			if err != nil {
+				return nil, err
+			}
+			res[k] = expression.NewLiteral(v, t)
+		default:
+			return nil, sql.ErrUnsupportedFeature.New(v.Type().String())
+		}
+	}
+	return res, nil
 }
 
 // QueryNodeWithBindings executes the query given with the bindings provided. If parsed is non-nil, it will be used
