+++ engine_test.go
@@ -1,0 +1,146 @@
+// Copyright 2023 Dolthub, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package sqle
+
+import (
+	"testing"
+	"time"
+
+	"github.com/dolthub/go-mysql-server/sql"
+	"github.com/dolthub/go-mysql-server/sql/expression"
+	"github.com/dolthub/go-mysql-server/sql/types"
+	"github.com/dolthub/vitess/go/vt/proto/query"
+	"github.com/stretchr/testify/require"
+)
+
+func TestBindingsToExprs(t *testing.T) {
+	type tc struct {
+		Name     string
+		Bindings map[string]*query.BindVariable
+		Result   map[string]sql.Expression
+		Err      bool
+	}
+
+	cases := []tc{
+		{
+			"Empty",
+			map[string]*query.BindVariable{},
+			map[string]sql.Expression{},
+			false,
+		},
+		{
+			"BadInt",
+			map[string]*query.BindVariable{
+				"v1": &query.BindVariable{Type: query.Type_INT8, Value: []byte("axqut")},
+			},
+			nil,
+			true,
+		},
+		{
+			"BadUint",
+			map[string]*query.BindVariable{
+				"v1": &query.BindVariable{Type: query.Type_UINT8, Value: []byte("-12")},
+			},
+			nil,
+			true,
+		},
+		{
+			"BadDecimal",
+			map[string]*query.BindVariable{
+				"v1": &query.BindVariable{Type: query.Type_DECIMAL, Value: []byte("axqut")},
+			},
+			nil,
+			true,
+		},
+		{
+			"BadBit",
+			map[string]*query.BindVariable{
+				"v1": &query.BindVariable{Type: query.Type_BIT, Value: []byte{byte(0), byte(0), byte(0), byte(0), byte(0), byte(0), byte(0), byte(0), byte(0)}},
+			},
+			nil,
+			true,
+		},
+		{
+			"BadDate",
+			map[string]*query.BindVariable{
+				"v1": &query.BindVariable{Type: query.Type_DATE, Value: []byte("00000000")},
+			},
+			nil,
+			true,
+		},
+		{
+			"BadYear",
+			map[string]*query.BindVariable{
+				"v1": &query.BindVariable{Type: query.Type_YEAR, Value: []byte("asdf")},
+			},
+			nil,
+			true,
+		},
+		{
+			"BadDatetime",
+			map[string]*query.BindVariable{
+				"v1": &query.BindVariable{Type: query.Type_DATETIME, Value: []byte("0000")},
+			},
+			nil,
+			true,
+		},
+		{
+			"BadTimestamp",
+			map[string]*query.BindVariable{
+				"v1": &query.BindVariable{Type: query.Type_TIMESTAMP, Value: []byte("0000")},
+			},
+			nil,
+			true,
+		},
+		{
+			"SomeTypes",
+			map[string]*query.BindVariable{
+				"i8":        &query.BindVariable{Type: query.Type_INT8, Value: []byte("12")},
+				"u64":       &query.BindVariable{Type: query.Type_UINT64, Value: []byte("4096")},
+				"bin":       &query.BindVariable{Type: query.Type_VARBINARY, Value: []byte{byte(0xC0), byte(0x00), byte(0x10)}},
+				"text":      &query.BindVariable{Type: query.Type_TEXT, Value: []byte("four score and seven years ago...")},
+				"bit":       &query.BindVariable{Type: query.Type_BIT, Value: []byte{byte(0x0f)}},
+				"date":      &query.BindVariable{Type: query.Type_DATE, Value: []byte("2020-10-20")},
+				"year":      &query.BindVariable{Type: query.Type_YEAR, Value: []byte("2020")},
+				"datetime":  &query.BindVariable{Type: query.Type_DATETIME, Value: []byte("2020-10-20T12:00:00Z")},
+				"timestamp": &query.BindVariable{Type: query.Type_TIMESTAMP, Value: []byte("2020-10-20T12:00:00Z")},
+			},
+			map[string]sql.Expression{
+				"i8":        expression.NewLiteral(int64(12), types.Int64),
+				"u64":       expression.NewLiteral(uint64(4096), types.Uint64),
+				"bin":       expression.NewLiteral([]byte{byte(0xC0), byte(0x00), byte(0x10)}, types.MustCreateBinary(query.Type_VARBINARY, int64(3))),
+				"text":      expression.NewLiteral("four score and seven years ago...", types.MustCreateStringWithDefaults(query.Type_TEXT, 33)),
+				"bit":       expression.NewLiteral(uint64(0x0f), types.MustCreateBitType(types.BitTypeMaxBits)),
+				"date":      expression.NewLiteral(time.Date(2020, time.Month(10), 20, 0, 0, 0, 0, time.UTC), types.Date),
+				"year":      expression.NewLiteral(int16(2020), types.Year),
+				"datetime":  expression.NewLiteral(time.Date(2020, time.Month(10), 20, 12, 0, 0, 0, time.UTC), types.Datetime),
+				"timestamp": expression.NewLiteral(time.Date(2020, time.Month(10), 20, 12, 0, 0, 0, time.UTC), types.Timestamp),
+			},
+			false,
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.Name, func(t *testing.T) {
+			res, err := bindingsToExprs(c.Bindings)
+			if !c.Err {
+				require.NoError(t, err)
+				require.Equal(t, c.Result, res)
+			} else {
+				require.Error(t, err, "%v", res)
+			}
+		})
+	}
+}
