+++ enginetest/queries/script_queries.go
@@ -17,6 +17,7 @@ package queries
 import (
 	"time"
 
+	"github.com/dolthub/vitess/go/vt/proto/query"
 	"gopkg.in/src-d/go-errors.v1"
 
 	gmstime "github.com/dolthub/go-mysql-server/internal/time"
@@ -76,6 +77,10 @@ type ScriptTestAssertion struct {
 
 	// Bindings are variable mappings only used for prepared tests
 	Bindings map[string]sql.Expression
+	
+	// VitessBindings are variable mappings only used for prepared tests, allowing us to more closely simulate the 
+	// server's wire path in engine tests
+	VitessBindings map[string]*query.BindVariable
 }
 
 // ScriptTests are a set of test scripts to run.
