+++ server/handler_test.go
@@ -33,7 +33,6 @@ import (
 	"github.com/dolthub/go-mysql-server/memory"
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/dolthub/go-mysql-server/sql/analyzer"
-	"github.com/dolthub/go-mysql-server/sql/expression"
 	"github.com/dolthub/go-mysql-server/sql/types"
 	"github.com/dolthub/go-mysql-server/sql/variables"
 )
@@ -928,126 +927,6 @@ func TestOkClosedConnection(t *testing.T) {
 		return nil
 	})
 	require.NoError(err)
-}
-
-func TestBindingsToExprs(t *testing.T) {
-	type tc struct {
-		Name     string
-		Bindings map[string]*query.BindVariable
-		Result   map[string]sql.Expression
-		Err      bool
-	}
-
-	cases := []tc{
-		{
-			"Empty",
-			map[string]*query.BindVariable{},
-			map[string]sql.Expression{},
-			false,
-		},
-		{
-			"BadInt",
-			map[string]*query.BindVariable{
-				"v1": &query.BindVariable{Type: query.Type_INT8, Value: []byte("axqut")},
-			},
-			nil,
-			true,
-		},
-		{
-			"BadUint",
-			map[string]*query.BindVariable{
-				"v1": &query.BindVariable{Type: query.Type_UINT8, Value: []byte("-12")},
-			},
-			nil,
-			true,
-		},
-		{
-			"BadDecimal",
-			map[string]*query.BindVariable{
-				"v1": &query.BindVariable{Type: query.Type_DECIMAL, Value: []byte("axqut")},
-			},
-			nil,
-			true,
-		},
-		{
-			"BadBit",
-			map[string]*query.BindVariable{
-				"v1": &query.BindVariable{Type: query.Type_BIT, Value: []byte{byte(0), byte(0), byte(0), byte(0), byte(0), byte(0), byte(0), byte(0), byte(0)}},
-			},
-			nil,
-			true,
-		},
-		{
-			"BadDate",
-			map[string]*query.BindVariable{
-				"v1": &query.BindVariable{Type: query.Type_DATE, Value: []byte("00000000")},
-			},
-			nil,
-			true,
-		},
-		{
-			"BadYear",
-			map[string]*query.BindVariable{
-				"v1": &query.BindVariable{Type: query.Type_YEAR, Value: []byte("asdf")},
-			},
-			nil,
-			true,
-		},
-		{
-			"BadDatetime",
-			map[string]*query.BindVariable{
-				"v1": &query.BindVariable{Type: query.Type_DATETIME, Value: []byte("0000")},
-			},
-			nil,
-			true,
-		},
-		{
-			"BadTimestamp",
-			map[string]*query.BindVariable{
-				"v1": &query.BindVariable{Type: query.Type_TIMESTAMP, Value: []byte("0000")},
-			},
-			nil,
-			true,
-		},
-		{
-			"SomeTypes",
-			map[string]*query.BindVariable{
-				"i8":        &query.BindVariable{Type: query.Type_INT8, Value: []byte("12")},
-				"u64":       &query.BindVariable{Type: query.Type_UINT64, Value: []byte("4096")},
-				"bin":       &query.BindVariable{Type: query.Type_VARBINARY, Value: []byte{byte(0xC0), byte(0x00), byte(0x10)}},
-				"text":      &query.BindVariable{Type: query.Type_TEXT, Value: []byte("four score and seven years ago...")},
-				"bit":       &query.BindVariable{Type: query.Type_BIT, Value: []byte{byte(0x0f)}},
-				"date":      &query.BindVariable{Type: query.Type_DATE, Value: []byte("2020-10-20")},
-				"year":      &query.BindVariable{Type: query.Type_YEAR, Value: []byte("2020")},
-				"datetime":  &query.BindVariable{Type: query.Type_DATETIME, Value: []byte("2020-10-20T12:00:00Z")},
-				"timestamp": &query.BindVariable{Type: query.Type_TIMESTAMP, Value: []byte("2020-10-20T12:00:00Z")},
-			},
-			map[string]sql.Expression{
-				"i8":        expression.NewLiteral(int64(12), types.Int64),
-				"u64":       expression.NewLiteral(uint64(4096), types.Uint64),
-				"bin":       expression.NewLiteral([]byte{byte(0xC0), byte(0x00), byte(0x10)}, types.MustCreateBinary(query.Type_VARBINARY, int64(3))),
-				"text":      expression.NewLiteral("four score and seven years ago...", types.MustCreateStringWithDefaults(query.Type_TEXT, 33)),
-				"bit":       expression.NewLiteral(uint64(0x0f), types.MustCreateBitType(types.BitTypeMaxBits)),
-				"date":      expression.NewLiteral(time.Date(2020, time.Month(10), 20, 0, 0, 0, 0, time.UTC), types.Date),
-				"year":      expression.NewLiteral(int16(2020), types.Year),
-				"datetime":  expression.NewLiteral(time.Date(2020, time.Month(10), 20, 12, 0, 0, 0, time.UTC), types.Datetime),
-				"timestamp": expression.NewLiteral(time.Date(2020, time.Month(10), 20, 12, 0, 0, 0, time.UTC), types.Timestamp),
-			},
-			false,
-		},
-	}
-
-	for _, c := range cases {
-		t.Run(c.Name, func(t *testing.T) {
-			res, err := bindingsToExprs(c.Bindings)
-			if !c.Err {
-				require.NoError(t, err)
-				require.Equal(t, c.Result, res)
-			} else {
-				require.Error(t, err, "%v", res)
-			}
-		})
-	}
 }
 
 // Tests the CLIENT_FOUND_ROWS capabilities flag
